garbage collect stale distributed locks
Under special conditions [1], cilium-agent could leak distributed locks
into the KVStore making all cilium-agents to deadlock. To avoid this
cilium-operator will garbage collect stale distributed locks. If a lock
is being held for more than 2x default value of cilium's lock lease TTL,
cilium-operator will delete that lock from the KVStore. On a first sight
this might not be reasonable but cilium-agents only use the distributed
locks to coordinated across each other which one should pick the
security identity for a new set of labels. Even if 2 cilium-agents pick
2 different security identities for the same set of labels, only the
first one will be able to publish that identity into the KVStore. The
2nd agent will fail to publish the security identity that it picked for
the set of labels:

[1]
1) connect to a etcd member (https://node-A:2379) to create a lock with
   a 25 second lease (the default for cilium)
2) drop all peer packets, not client packets, to and from that etcd
   member (https://node-A:2380).
3) try releasing that lock immediately. The unlock will fail with the
   error level=error msg="Unable to unlock kvstore lock"
   error="etcdserver: request timed out".
4) As soon the unlock fails, resume connectivity dropped in 2), this
   re-connectivity needs to happen before the lease expires, which means
   the Unlock on step 3) needs to fail in less than 25 seconds, this is
   what happen for my case.
5) since the client resumed connectivity with the server in less than 25
   seconds, the lease didn't expire which means the server still thinks
   the client is still holding the lock but the client assumes the lock was
   unlocked.

Signed-off-by: André Martins <andre@cilium.io>